Fix live activity implementation
