+++ HaukWidgetsExtension/LocationSharingLiveActivity.swift
@@ -5,33 +5,66 @@ import SwiftUI
 struct LocationSharingLiveActivity: Widget {
     var body: some WidgetConfiguration {
         ActivityConfiguration(for: LocationSharingAttributes.self) { context in
-            VStack(alignment: .leading) {
-                Text("Sharing location")
-                    .font(.headline)
-                if let location = context.state.lastLocation {
-                    Text(location)
-                        .font(.caption)
+            HStack(spacing: 12) {
+                Image("AppLogo")
+                    .resizable()
+                    .frame(width: 36, height: 36)
+                VStack(alignment: .leading) {
+                    Text("Sharing location")
+                        .font(.headline)
+                    Text(context.state.expiryDate, style: .timer)
+                        .font(.system(.body, design: .monospaced))
                 }
+                Spacer()
+                Button("Stop") {
+                    Task {
+                        await SharingManager.shared.stopSharing()
+                    }
+                }
+                .buttonStyle(.borderedProminent)
             }
             .padding()
         } dynamicIsland: { context in
             DynamicIsland {
                 DynamicIslandExpandedRegion(.leading) {
-                    Text("Hauk")
+                    Image("AppLogo")
+                        .resizable()
+                        .frame(width: 24, height: 24)
                 }
                 DynamicIslandExpandedRegion(.trailing) {
-                    if let location = context.state.lastLocation {
-                        Text(location)
+                    Button("Stop") {
+                        Task {
+                            await SharingManager.shared.stopSharing()
+                        }
                     }
+                    .buttonStyle(.bordered)
+                }
+                DynamicIslandExpandedRegion(.center) {
+                    Text(context.state.expiryDate, style: .timer)
+                        .monospacedDigit()
                 }
             } compactLeading: {
-                Image(systemName: "location")
+                CountdownRingView(expiryDate: context.state.expiryDate,
+                                  startDate: context.attributes.startDate)
+                    .frame(width: 22, height: 22)
             } compactTrailing: {
-                Text("Hauk")
+                Image(systemName: "location.fill")
             } minimal: {
-                Image(systemName: "location")
+                CountdownRingView(expiryDate: context.state.expiryDate,
+                                  startDate: context.attributes.startDate)
+                    .frame(width: 22, height: 22)
             }
         }
+    }
+}
+
+struct CountdownRingView: View {
+    let expiryDate: Date
+    let startDate: Date
+
+    var body: some View {
+        ProgressView(timerInterval: startDate...expiryDate, countsDown: true)
+            .progressViewStyle(.circular)
     }
 }
